pyproject.toml
[project]
name = "parking-spot-monitoring"
version = "1.0.0"
description = "Monitor parking spot availability using UniFi Protect cameras and YOLOv8"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "uiprotect>=6.0.0",
    "ultralytics>=8.1.0",
    "opencv-python>=4.9.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
